requirements.txt
absl-py==1.0.0
altgraph==0.17.2
apex @ file:///C:/apex
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.4.0
black==22.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
cryptography==36.0.1
cxxfilt==0.3.0
cycler==0.11.0
dataclasses==0.8
defusedxml==0.7.1
Django==2.2.11
django-allauth==0.41.0
django-environ==0.4.5
django-filter==2.2.0
django-jsonfield==1.4.0
django-querysetsequence==0.12
django-rest-auth==0.9.5
django-spa==0.3.5
djangorestframework==3.11.0
djangorestframework-csv==2.1.0
djitellopy==2.4.0
docker-pycreds==0.4.0
drawnow==0.72.5
drf-nested-routers==0.91
ffmpeg==1.4
filetype==1.0.5
future==0.18.2
gast==0.3.3
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==2.10.0
idna==2.8
ijson==2.6.1
importlib-metadata==4.8.3
imutils==0.5.4
iniconfig==1.1.1
inputs==0.5
iorodeo-potentiostat==0.0.4
iso8601==0.1.12
Jinja2==3.0.3
jsonschema==3.2.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
labelImg==1.8.6
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
messagebox==0.1.0
mypy-extensions==0.4.3
numpy==1.18.5
nvidia-cuda-runtime-cu11==2021.12.20
nvidia-cuda-runtime-cu116==11.6.55
nvidia-pyindex==1.0.9
oauthlib==3.2.0
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pathspec==0.9.0
pathtools==0.1.2
pefile==2021.9.3
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
progressbar==2.5
progressbar2==3.55.0
progressbar33==2.4
promise==2.3
protobuf==3.19.4
psutil==5.6.7
psycopg2==2.8.6
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.2
pyinstaller==4.8
pyinstaller-hooks-contrib==2022.0
pyparsing==3.0.7
pyproject==1.3.1
pyproject-toml==0.0.10
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyqt5-tools==5.15.4.3.2
pyrsistent==0.18.0
pyserial==3.5
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.6
python-utils==3.1.0
python3-openid==3.2.0
pytz==2021.3
pyvips==2.1.8
pywin32-ctypes==0.2.0
PyYAML==6.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
remo==0.6.1
remo-python==0.1.14
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.5.4
seaborn==0.11.2
sendgrid==6.0.5
sentry-sdk==1.5.5
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
thop==0.0.31.post2005241907
tk==0.1.0
toml==0.10.2
tomli==1.2.3
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
typed-ast==1.5.2
typer==0.1.1
typing_extensions==4.0.1
unicodecsv==0.14.1
urllib3==1.25.11
waitress==1.4.3
wandb==0.12.10
Werkzeug==2.0.3
whitenoise==5.0.1
wrapt==1.13.3
yaspin==2.1.0
zipp==3.6.0
